Move component definition to separate files.
