+++ website/src/components/HomepageFeatures/featureGallery.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import styles from './styles.module.css';
+
+function FeatureGallery({ header, images, description }) {
+  return (
+    <div className={styles.featureGallery}>
+      <h2 className={styles.header}>{header}</h2>
+
+      <div className={styles.imageGallery}>
+        {images.map((image, index) => (
+          <a key={index} href={image.link} target="_blank" rel="noopener noreferrer">
+            <img className={styles.galleryImage} src={image.src} alt={image.alt || `Image ${index + 1}`} />
+          </a>
+        ))}
+      </div>
+
+      {description && <p className={styles.description}>{description}</p>}
+    </div>
+  );
+}
+
+export default FeatureGallery;
